requirements.txt
flask
redis
pillow
pylast
requests